chat: rename githubTarballFetcher to githubRepoFetcher
The file no longer contains tarball logic — it fetches the repo
tree and individual blobs via api.github.com. Rename to match.
